Moved atoms display from Feed page to Home page
